Update Feedback icon.
